Fix nondeterministic Ambiguous error messages
An earlier commit made the desugaring process deterministic. This PR switches
another HashMap in config.rs to use IndexMap, though I doubt it actually makes a
difference for that particular structure.
More importantly, the way we create the Ambiguous error message in one case now
involves iterating through the values of a child subtrie, which is unfortunately
not deterministic. In order to handle this, we must pay the cost of a sort.
Thankfully, this only happens at error time so it should be whatever.