Implement the help view
Adds a help view to the app when the user presses the '?' key. When the user does press this key, it will update the view to include annotation messages to the right of each line explaining each one. Furthermore, it will append a help section with some useful tips and a pointer to the README of the project. A plaintext example is copy-pasted below:

```
 a ⤙ ...    (Hit 'a' to continue the current prefix in the tree.)
 b ⇀ 'bar'    (Hit 'b' to execute the 'bar' target.)
 f ⇀ 'foo'    (Hit 'f' to execute the 'foo' target.)
 n ⤙ ...    (Hit 'n' to continue the current prefix in the tree.)

 prefix: ''    (This is the current prefix. You've pressed '' so far.)
 ? - toggle help    (Toggles this annotated view!)
 ••    (Just a waiting animation.)

 Some helpful things ===
 • Press '.' to execute the current prefix, assuming it points to something executable.
 • Press <backspace> to undo a character you've pressed and go back up the tree.
 • Press <Control-C>/<Esc> to exit Jam :(.
 • READ the README.md (pls)!
```

Note that the UI currently does not handle a large number of targets very well. This approach to the HELP view may have to get creative when we update the UI for a larger number of keys. For example, we may want to implement paging behavior so this help view concept still works without much change, or we may need to overhaul it completely if we decide to have columns of keys. We'll have to see, but for now I think its OK.