Fix issue where last help section line was not being printed
This seemed to be caused by our original value of VIEWPORT_BASE_HEIGHT not being
high enough (off by one). While doing this, I also did a bit of a refactor, such
that VIEWPORT_BASE_HEIGHT is no longer hardcoded to a specific number, and
instead, is set to the sum of a bunch of other consts, that are used to set the
sizes of the other sections. This should make it so that changes to ui.rs won't
inadvertently trigger this bug to occur again.