Make jam dumb (#2)
* Make demo.yml contain a multi-line command

Mostly for coverage I suppose.

* Move Target/Shortcut to new store.ts

* Move additional types to store.rs

I am starting to think it might be better to keep these types in
jam.rs, since I think the main value add of this work is to decouple
the store implementation from the parsing/execution implementation.

* Update the signatures for TargetStore to match jam.rs

* Add constructor to trait

* Add boilerplate for TrieDagStore impl

* Bring over ParseResult and signature of new()

* Implement new()

* Implement mappings

* Implement next()

* Remove the execute methods

These actually don't make sense existing in a collection struct.

* Implement lookup()

* Update jam to use store.rs

* Bump rust + time version

The time bump had to happen cause the version of time we were using
was incompatible with the newer rust version.

* Get rid of warnings in both jam.rs and store.rs

* Remove unnecessary pub

* Update reconciler to function off of TargetStore trait rather than ShortcutTrie

* First pass at SimpleStore

* Fix bug where .first() was being used in place of len() == 1

* Fix issue where looking up the prefix directly would always give false

This is cause we now have a map of shortcuts -> targets, so the prefix
will usually not exist since it will never be an exact match. We have
to check if any of the shortcuts contain it as a prefix.

* Fix bug where non-branch, non-leaf nodes weren't rendered as BranchKey

The message of this commit should clue you in that maybe something is
wrong. See the diff, I write a comment to that effect.

* Use SimpleStore in store.rs tests

* Delete TrieDagStore

See ya, won't miss ya.

* Add error detection to SimpleStore for non-cycle cases

* Remove auxiliary types and helpers for TrieDagStore

Some things I missed in the earlier commit. Also updated a variable
name that was using camel case when it shouldn't be.

* Implement cycle detection

* [wip] adding concept of parent key

* Add support for a log filepath

Logging to stderr is just too annoying with how easily it can get borked.

* Simplify and fix next_no_conflict()